+++ jsmooch-lib/src/system/nes/nes_cart.c
@@ -137,7 +137,7 @@ void read_ROM_RAM(struct NES_cart* this, char* inp, size_t inp_size)
 
 }
 
-u32 NES_cart_load_ROM_from_RAM(struct NES_cart* this, char* fil, size_t fil_sz)
+u32 NES_cart_load_ROM_from_RAM(struct NES_cart* this, char* fil, u64 fil_sz)
 {
     if ((fil[0] != 0x4E) || (fil[1] != 0x45) ||
         (fil[2] != 0x53) || (fil[3] != 0x1A)) {
